chore: update FilterIcon component and RankingTable component for admin page
